refactor: move caption schema to shared folder
