+++ src/nest/captions/captions.module.ts
@@ -2,7 +2,7 @@ import { Module } from "@nestjs/common";
 import { MongooseModule } from "@nestjs/mongoose";
 import { CaptionsController } from "./captions.controller";
 import { CaptionsService } from "./captions.service";
-import { Caption, CaptionSchema } from "./schemas/caption.schema";
+import { Caption, CaptionSchema } from "../shared/schemas/caption.schema";
 import { Video, VideoSchema } from "./schemas/video.schema";
 import { Captioner, CaptionerSchema } from "./schemas/captioner.schema";
